FASTAPI/requirements.txt
fastapi
pygame
uvicorn
mysql-connector-python 
python-dotenv 
msal
pydantic
python-jose
passlib
snowflake